fix: удалён UserTest, добавлены UserRegisterTest и UserLoginTest
